new files are added

+++ mac/shop/models.py
@@ -7,7 +7,7 @@ class Product(models.Model):
     product_name = models.CharField(max_length = 50)
     category = models.CharField(max_length = 50, default = "")
     sub_category = models.CharField(max_length = 50, default = "")
-    price = models.IntegerField(default = 0)
+    price = models.FloatField(default = 0)
     desc = models.CharField(max_length = 500)
     pub_date = models.DateField()
     image = models.ImageField(upload_to = "shop/images", default = "")
